Refactoring naming convention
