+++ aws-dotnet-rest-api-with-dynamodb/DotNetServerless.sln
@@ -1,0 +1,71 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26124.0
+MinimumVisualStudioVersion = 15.0.26124.0
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{E45346EE-74B7-4F5B-943C-FEFDE57124D0}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{DBC29D13-84FE-4A90-B785-E325BDD494A8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DotNetServerless.Domain", "src\DotNetServerless.Domain\DotNetServerless.Domain.csproj", "{EE96C435-FFE9-4F98-9865-47B0DC7C2427}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DotNetServerless.Functions", "src\DotNetServerless.Functions\DotNetServerless.Functions.csproj", "{FB087A82-65A5-439C-A344-63FC8F95FB71}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DotNetServerless.Functions.Tests", "tests\DotNetServerless.Functions.Tests\DotNetServerless.Functions.Tests.csproj", "{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{EE96C435-FFE9-4F98-9865-47B0DC7C2427}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EE96C435-FFE9-4F98-9865-47B0DC7C2427}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EE96C435-FFE9-4F98-9865-47B0DC7C2427}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{EE96C435-FFE9-4F98-9865-47B0DC7C2427}.Debug|x64.Build.0 = Debug|Any CPU
+		{EE96C435-FFE9-4F98-9865-47B0DC7C2427}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{EE96C435-FFE9-4F98-9865-47B0DC7C2427}.Debug|x86.Build.0 = Debug|Any CPU
+		{EE96C435-FFE9-4F98-9865-47B0DC7C2427}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EE96C435-FFE9-4F98-9865-47B0DC7C2427}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EE96C435-FFE9-4F98-9865-47B0DC7C2427}.Release|x64.ActiveCfg = Release|Any CPU
+		{EE96C435-FFE9-4F98-9865-47B0DC7C2427}.Release|x64.Build.0 = Release|Any CPU
+		{EE96C435-FFE9-4F98-9865-47B0DC7C2427}.Release|x86.ActiveCfg = Release|Any CPU
+		{EE96C435-FFE9-4F98-9865-47B0DC7C2427}.Release|x86.Build.0 = Release|Any CPU
+		{FB087A82-65A5-439C-A344-63FC8F95FB71}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FB087A82-65A5-439C-A344-63FC8F95FB71}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FB087A82-65A5-439C-A344-63FC8F95FB71}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{FB087A82-65A5-439C-A344-63FC8F95FB71}.Debug|x64.Build.0 = Debug|Any CPU
+		{FB087A82-65A5-439C-A344-63FC8F95FB71}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{FB087A82-65A5-439C-A344-63FC8F95FB71}.Debug|x86.Build.0 = Debug|Any CPU
+		{FB087A82-65A5-439C-A344-63FC8F95FB71}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FB087A82-65A5-439C-A344-63FC8F95FB71}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FB087A82-65A5-439C-A344-63FC8F95FB71}.Release|x64.ActiveCfg = Release|Any CPU
+		{FB087A82-65A5-439C-A344-63FC8F95FB71}.Release|x64.Build.0 = Release|Any CPU
+		{FB087A82-65A5-439C-A344-63FC8F95FB71}.Release|x86.ActiveCfg = Release|Any CPU
+		{FB087A82-65A5-439C-A344-63FC8F95FB71}.Release|x86.Build.0 = Release|Any CPU
+		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}.Debug|x64.Build.0 = Debug|Any CPU
+		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}.Debug|x86.Build.0 = Debug|Any CPU
+		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}.Release|x64.ActiveCfg = Release|Any CPU
+		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}.Release|x64.Build.0 = Release|Any CPU
+		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}.Release|x86.ActiveCfg = Release|Any CPU
+		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{EE96C435-FFE9-4F98-9865-47B0DC7C2427} = {E45346EE-74B7-4F5B-943C-FEFDE57124D0}
+		{FB087A82-65A5-439C-A344-63FC8F95FB71} = {E45346EE-74B7-4F5B-943C-FEFDE57124D0}
+		{FED148BE-1C5D-41E2-8D93-0EE892A09CC2} = {DBC29D13-84FE-4A90-B785-E325BDD494A8}
+	EndGlobalSection
+EndGlobal
