--- aws-dotnet-rest-api-with-dynamodb/TodoLambda.sln
@@ -1,54 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26124.0
-MinimumVisualStudioVersion = 15.0.26124.0
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{E45346EE-74B7-4F5B-943C-FEFDE57124D0}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TodoLambda.Domain", "src\TodoLambda.Domain\TodoLambda.Domain.csproj", "{7688CD59-5F95-4CCA-8B6E-0F58F9DC35D4}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TodoLambda.Functions", "src\TodoLambda.Functions\TodoLambda.Functions.csproj", "{62C1880A-F35D-4D01-8C9C-9D53E5D7C1EB}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|Any CPU = Release|Any CPU
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7688CD59-5F95-4CCA-8B6E-0F58F9DC35D4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7688CD59-5F95-4CCA-8B6E-0F58F9DC35D4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7688CD59-5F95-4CCA-8B6E-0F58F9DC35D4}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{7688CD59-5F95-4CCA-8B6E-0F58F9DC35D4}.Debug|x64.Build.0 = Debug|Any CPU
-		{7688CD59-5F95-4CCA-8B6E-0F58F9DC35D4}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{7688CD59-5F95-4CCA-8B6E-0F58F9DC35D4}.Debug|x86.Build.0 = Debug|Any CPU
-		{7688CD59-5F95-4CCA-8B6E-0F58F9DC35D4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7688CD59-5F95-4CCA-8B6E-0F58F9DC35D4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7688CD59-5F95-4CCA-8B6E-0F58F9DC35D4}.Release|x64.ActiveCfg = Release|Any CPU
-		{7688CD59-5F95-4CCA-8B6E-0F58F9DC35D4}.Release|x64.Build.0 = Release|Any CPU
-		{7688CD59-5F95-4CCA-8B6E-0F58F9DC35D4}.Release|x86.ActiveCfg = Release|Any CPU
-		{7688CD59-5F95-4CCA-8B6E-0F58F9DC35D4}.Release|x86.Build.0 = Release|Any CPU
-		{62C1880A-F35D-4D01-8C9C-9D53E5D7C1EB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{62C1880A-F35D-4D01-8C9C-9D53E5D7C1EB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{62C1880A-F35D-4D01-8C9C-9D53E5D7C1EB}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{62C1880A-F35D-4D01-8C9C-9D53E5D7C1EB}.Debug|x64.Build.0 = Debug|Any CPU
-		{62C1880A-F35D-4D01-8C9C-9D53E5D7C1EB}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{62C1880A-F35D-4D01-8C9C-9D53E5D7C1EB}.Debug|x86.Build.0 = Debug|Any CPU
-		{62C1880A-F35D-4D01-8C9C-9D53E5D7C1EB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{62C1880A-F35D-4D01-8C9C-9D53E5D7C1EB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{62C1880A-F35D-4D01-8C9C-9D53E5D7C1EB}.Release|x64.ActiveCfg = Release|Any CPU
-		{62C1880A-F35D-4D01-8C9C-9D53E5D7C1EB}.Release|x64.Build.0 = Release|Any CPU
-		{62C1880A-F35D-4D01-8C9C-9D53E5D7C1EB}.Release|x86.ActiveCfg = Release|Any CPU
-		{62C1880A-F35D-4D01-8C9C-9D53E5D7C1EB}.Release|x86.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{7688CD59-5F95-4CCA-8B6E-0F58F9DC35D4} = {E45346EE-74B7-4F5B-943C-FEFDE57124D0}
-		{62C1880A-F35D-4D01-8C9C-9D53E5D7C1EB} = {E45346EE-74B7-4F5B-943C-FEFDE57124D0}
-	EndGlobalSection
-EndGlobal
