+++ aws-dotnet-rest-api-with-dynamodb/src/DotNetServerless.Domain/DotNetServerless.Domain.csproj
@@ -1,0 +1,16 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="AWSSDK.DynamoDBv2" Version="3.3.14.12" />
+    <PackageReference Include="MediatR" Version="5.1.0" />
+    <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="5.1.0" />
+    <PackageReference Include="AWSSDK.Core" Version="3.3.29.3" />
+    <PackageReference Include="AWSSDK.Core" Version="3.3.29.3" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="2.1.1.0" />
+    <PackageReference Include="Newtonsoft.Json" Version="11.0.1" />
+  </ItemGroup>
+</Project>
