+++ aws-dotnet-rest-api-with-dynamodb/src/DotNetServerless.Domain/Entity/Item.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace DotNetServerless.Domain.Entity
+{
+  public class Item
+  {
+    public Guid Id { get; set; }
+    public string Description { get; set; }
+    public bool IsChecked { get; set; }
+  }
+}
