+++ aws-dotnet-rest-api-with-dynamodb/src/DotNetServerless.Domain/Infrastructure/Configs/DynamoDbConfiguration.cs
@@ -1,0 +1,12 @@
+namespace DotNetServerless.Domain.Infrastructure.Configs
+{
+  public interface IDynamoDbConfiguration
+  {
+    string TableName { get; set; }
+  }
+
+  public class DynamoDbConfiguration : IDynamoDbConfiguration
+  {
+    public string TableName { get; set; }
+  }
+}
