+++ aws-dotnet-rest-api-with-dynamodb/src/DotNetServerless.Domain/Requests/CreateItemRequest.cs
@@ -1,0 +1,20 @@
+using DotNetServerless.Domain.Entity;
+using MediatR;
+
+namespace DotNetServerless.Domain.Requests
+{
+  public class CreateItemRequest : IRequest<Item>
+  {
+    public string Description { get; set; }
+    public bool IsChecked { get; set; }
+
+    public Item Map()
+    {
+      return new Item
+      {
+        Description = Description,
+        IsChecked = IsChecked
+      };
+    }
+  }
+}
