+++ aws-dotnet-rest-api-with-dynamodb/src/DotNetServerless.Domain/Requests/GetItemRequest.cs
@@ -1,0 +1,11 @@
+using System;
+using DotNetServerless.Domain.Entity;
+using MediatR;
+
+namespace DotNetServerless.Domain.Requests
+{
+  public class GetItemRequest : IRequest<Item>
+  {
+    public Guid Id { get; set; }
+  }
+}
