+++ aws-dotnet-rest-api-with-dynamodb/src/DotNetServerless.Domain/Responses/ItemResponse.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace DotNetServerless.Domain.Responses
+{
+  public class ItemResponse
+  {
+    public Guid Id { get; set; }
+    public string Description { get; set; }
+    public bool IsChecked { get; set; }
+  }
+}
