+++ aws-dotnet-rest-api-with-dynamodb/src/DotNetServerless.Functions/Extensions/ServicesExtensions.cs
@@ -1,0 +1,18 @@
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace DotNetServerless.Functions.Extensions
+{
+  public static class ServiceCollectionExtensions
+  {
+    public static IServiceCollection BindAndConfigure<TConfig>(this IServiceCollection services,
+      IConfigurationSection section, TConfig config)
+      where TConfig : class, new()
+    {
+      section.Bind(config);
+      services.AddSingleton(config);
+
+      return services;
+    }
+  }
+}
