+++ aws-dotnet-rest-api-with-dynamodb/src/DotNetServerless.Functions/Lambda/CreateFunction.cs
@@ -1,0 +1,32 @@
+using System;
+using System.Threading.Tasks;
+using Amazon.Lambda.Core;
+using Amazon.Lambda.Serialization.Json;
+using DotNetServerless.Domain.Entity;
+using DotNetServerless.Domain.Requests;
+using MediatR;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace DotNetServerless.Functions.Lambda
+{
+  public class CreateFunction
+  {
+    private readonly IServiceProvider _serviceProvider;
+
+    public CreateFunction() : this(Startup.BuildContainer())
+    {
+    }
+
+    public CreateFunction(IServiceProvider serviceProvider)
+    {
+      _serviceProvider = serviceProvider;
+    }
+
+    [LambdaSerializer(typeof(JsonSerializer))]
+    public async Task<Item> Run(CreateItemRequest request)
+    {
+      var mediator = _serviceProvider.GetService<IMediator>();
+      return await mediator.Send(request);
+    }
+  }
+}
