+++ aws-dotnet-rest-api-with-dynamodb/src/DotNetServerless.Functions/Program.cs
@@ -1,0 +1,9 @@
+﻿namespace DotNetServerless.Functions
+{
+  public class Program
+  {
+    public static void Main(string[] args)
+    {
+    }
+  }
+}
