+++ aws-dotnet-rest-api-with-dynamodb/src/DotNetServerless.Functions/build.sh
