+++ aws-dotnet-rest-api-with-dynamodb/src/DotNetServerless.Functions/serverless.yml
