--- aws-dotnet-rest-api-with-dynamodb/src/TodoLambda.Domain/Entity/Item.cs
@@ -1,11 +1,0 @@
-using System;
-
-namespace TodoLambda.Domain.Entity
-{
-  public class Item
-  {
-    public Guid Id { get; set; }
-    public string Description { get; set; }
-    public bool IsChecked { get; set; }
-  }
-}
