--- aws-dotnet-rest-api-with-dynamodb/src/TodoLambda.Domain/Infrastructure/Configs/DynamoDbConfiguration.cs
@@ -1,12 +1,0 @@
-namespace TodoLambda.Domain.Infrastructure.Configs
-{
-  public interface IDynamoDbConfiguration
-  {
-    string TableName { get; set; }
-  }
-
-  public class DynamoDbConfiguration : IDynamoDbConfiguration
-  {
-    public string TableName { get; set; }
-  }
-}
