--- aws-dotnet-rest-api-with-dynamodb/src/TodoLambda.Domain/Infrastructure/Repositories/IItemRepository.cs
@@ -1,13 +1,0 @@
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using TodoLambda.Domain.Entity;
-
-namespace TodoLambda.Domain.Infrastructure.Repositories
-{
-  public interface IItemRepository
-  {
-    Task Save(Item item);
-
-    Task<IEnumerable<Item>> GetById(string id);
-  }
-}
