--- aws-dotnet-rest-api-with-dynamodb/src/TodoLambda.Domain/Requests/CreateItemRequest.cs
@@ -1,20 +1,0 @@
-using MediatR;
-using TodoLambda.Domain.Entity;
-
-namespace TodoLambda.Domain.Requests
-{
-  public class CreateItemRequest : IRequest<Item>
-  {
-    public string Description { get; set; }
-    public bool IsChecked { get; set; }
-
-    public Item Map()
-    {
-      return new Item
-      {
-        Description = Description,
-        IsChecked = IsChecked
-      };
-    }
-  }
-}
