--- aws-dotnet-rest-api-with-dynamodb/src/TodoLambda.Domain/Requests/GetItemRequest.cs
@@ -1,11 +1,0 @@
-using System;
-using MediatR;
-using TodoLambda.Domain.Entity;
-
-namespace TodoLambda.Domain.Requests
-{
-  public class GetItemRequest : IRequest<Item>
-  {
-    public Guid Id { get; set; }
-  }
-}
