--- aws-dotnet-rest-api-with-dynamodb/src/TodoLambda.Domain/Requests/UpdateItemRequest.cs
@@ -1,23 +1,0 @@
-using System;
-using MediatR;
-using TodoLambda.Domain.Entity;
-
-namespace TodoLambda.Domain.Requests
-{
-  public class UpdateItemRequest : IRequest<Item>
-  {
-    public Guid Id { get; set; }
-    public string Description { get; set; }
-    public bool IsChecked { get; set; }
-
-    public Item Map()
-    {
-      return new Item
-      {
-        Id = Id,
-        Description = Description,
-        IsChecked = IsChecked
-      };
-    }
-  }
-}
