--- aws-dotnet-rest-api-with-dynamodb/src/TodoLambda.Domain/Responses/ItemResponse.cs
@@ -1,11 +1,0 @@
-using System;
-
-namespace TodoLambda.Domain.Responses
-{
-  public class ItemResponse
-  {
-    public Guid Id { get; set; }
-    public string Description { get; set; }
-    public bool IsChecked { get; set; }
-  }
-}
