--- aws-dotnet-rest-api-with-dynamodb/src/TodoLambda.Functions/Extensions/ServicesExtensions.cs
@@ -1,18 +1,0 @@
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace TodoLambda.Functions.Extensions
-{
-  public static class ServiceCollectionExtensions
-  {
-    public static IServiceCollection BindAndConfigure<TConfig>(this IServiceCollection services,
-      IConfigurationSection section, TConfig config)
-      where TConfig : class, new()
-    {
-      section.Bind(config);
-      services.AddSingleton(config);
-
-      return services;
-    }
-  }
-}
