--- aws-dotnet-rest-api-with-dynamodb/src/TodoLambda.Functions/Program.cs
@@ -1,9 +1,0 @@
-﻿namespace TodoLambda.Functions
-{
-  public class Program
-  {
-    public static void Main(string[] args)
-    {
-    }
-  }
-}
