--- aws-dotnet-rest-api-with-dynamodb/src/TodoLambda.Functions/Startup.cs
@@ -1,39 +1,0 @@
-﻿using System;
-using System.IO;
-using MediatR;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using TodoLambda.Domain.Infrastructure;
-using TodoLambda.Domain.Infrastructure.Configs;
-using TodoLambda.Domain.Infrastructure.Repositories;
-using TodoLambda.Functions.Extensions;
-
-namespace TodoLambda.Functions
-{
-  public class Startup
-  {
-    public static IServiceProvider BuildContainer()
-    {
-      var configuration = new ConfigurationBuilder()
-        .SetBasePath(Directory.GetCurrentDirectory())
-        .AddEnvironmentVariables()
-        .Build();
-
-      return ConfigureServices(configuration);
-    }
-
-
-    private static IServiceProvider ConfigureServices(IConfigurationRoot configurationRoot)
-    {
-      var services = new ServiceCollection();
-
-      services
-        .AddMediatR()
-        .AddTransient(typeof(IAwsClientFactory<>), typeof(AwsClientFactory<>))
-        .AddTransient<IItemRepository, ItemDynamoRepository>()
-        .BindAndConfigure(configurationRoot.GetSection("DYNAMODB_TABLE"), new DynamoDbConfiguration());
-
-      return services.BuildServiceProvider();
-    }
-  }
-}
